Re-enable JSON alongside JSONRPC
This introduces a combinator `UsingContentType` that allows client code
to explicitly specify the content type they would like. This is
backwards compatible: unless explicitly specified, the default is still
`JSONRPC`, like it is now.

`UsingContentType` also has its own `HasClient` instance for `(:<|>)`;
this makes it possible to define an API without specifying the content
type, and then at the top-level wrap the entire thing in a single call
to `UsingContentType`.